Generalise and reuse Relationship code
Now we can use the code created for Applicants in the Respondent steps.
DRY views and i18n as a consequence of this.
