+++ app/controllers/steps/respondent/relationship_controller.rb
@@ -1,16 +1,7 @@
 module Steps
   module Respondent
     class RelationshipController < Steps::RespondentStepController
-      def edit
-        @form_object = RelationshipForm.new(
-          c100_application: current_c100_application,
-          relationship: current_c100_application.relationship
-        )
-      end
-
-      def update
-        update_and_advance(RelationshipForm)
-      end
+      include RelationshipStep
     end
   end
 end
