+++ app/views/steps/respondent/relationship/edit.html.erb
@@ -4,13 +4,6 @@
   <div class="column-two-thirds">
     <%= step_header %>
 
-    <h1 class="heading-xlarge gv-u-heading-xxlarge"><%=t '.heading' %></h1>
-
-    <p class="lede gv-u-text-lede"><%=t '.lead_text' %></p>
-
-    <%= step_form @form_object do |f| %>
-      <!-- TODO: Add form fields here -->
-      <%= f.submit class: 'button' %>
-    <% end %>
+    <%= render partial: 'steps/shared/relationship_step', locals: { form_object: @form_object } %>
   </div>
 </div>
