+++ spec/controllers/steps/applicant/relationship_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 RSpec.describe Steps::Applicant::RelationshipController, type: :controller do
-  it_behaves_like 'a relationship step controller', Steps::Applicant::RelationshipForm, C100App::ApplicantDecisionTree
+  it_behaves_like 'a relationship step controller', Steps::Shared::RelationshipForm, C100App::ApplicantDecisionTree
 end
