+++ spec/controllers/steps/respondent/relationship_controller_spec.rb
@@ -1,5 +1,5 @@
 require 'rails_helper'
 
 RSpec.describe Steps::Respondent::RelationshipController, type: :controller do
-  it_behaves_like 'an intermediate step controller', Steps::Respondent::RelationshipForm, C100App::RespondentDecisionTree
+  it_behaves_like 'a relationship step controller', Steps::Shared::RelationshipForm, C100App::RespondentDecisionTree
 end
