+++ spec/support/relationship_step_controller_shared_examples.rb
@@ -1,13 +1,9 @@
 require 'rails_helper'
 
-# TODO: at the moment this `shared examples` are not really generic enough to cope with
-# models other than `Applicant`, but once we start implementing the relationships for another
-# model, for example `Respondent`, it will be more clear what changes we need to make.
-#
 RSpec.shared_examples 'a relationship step controller' do |form_class, decision_tree_class|
   let(:relationships) { double('relationships').as_null_object }
   let(:children)      { double('children').as_null_object }
-  let(:applicant)     { double('Applicant') }
+  let(:person)        { double('Person') }
   let(:child)         { double('Child') }
 
   let(:existing_case) { instance_double(C100Application, relationships: relationships, children: children) }
@@ -40,10 +36,10 @@ RSpec.shared_examples 'a relationship step controller' do |form_class, decision_
     context 'when a case in progress is in the session' do
       before do
         allow(form_class).to receive(:new).and_return(form_object)
-        allow(controller).to receive(:current_record).and_return(applicant)
+        allow(controller).to receive(:current_record).and_return(person)
         allow(children).to receive(:find).with('123').and_return(child)
 
-        expect(relationships).to receive(:find_or_initialize_by).with(applicant: applicant, child: child)
+        expect(relationships).to receive(:find_or_initialize_by).with(person: person, child: child)
       end
 
       context 'when the form saves successfully' do
@@ -89,9 +85,9 @@ RSpec.shared_examples 'a relationship step controller' do |form_class, decision_
 
     context 'when a case exists in the session' do
       before do
-        allow(controller).to receive(:current_record).and_return(applicant)
+        allow(controller).to receive(:current_record).and_return(person)
         allow(children).to receive(:find).with('123').and_return(child)
-        expect(relationships).to receive(:find_or_initialize_by).with(applicant: applicant, child: child)
+        expect(relationships).to receive(:find_or_initialize_by).with(person: person, child: child)
       end
 
       it 'responds with HTTP success' do
